server: move inMemoryResponseWriter to handler.go
